+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CommonDbContext.cs
@@ -135,7 +135,7 @@ internal abstract class CommonDbContext : DbContext
 
     public DbSet<PackageSchemaHistory> PackageSchemaHistory => Set<PackageSchemaHistory>();
 
-    public DbSet<ComponentSchema> ComponentSchema => Set<ComponentSchema>();
+    public DbSet<ValidatorEmissions> ValidatorEmissions => Set<ValidatorEmissions>();
 
     public CommonDbContext(DbContextOptions options)
         : base(options)
@@ -371,5 +371,8 @@ internal abstract class CommonDbContext : DbContext
 
         modelBuilder.Entity<ValidatorActiveSetHistory>()
             .HasIndex(e => e.Epoch);
+
+        modelBuilder.Entity<ValidatorEmissions>()
+            .HasIndex(e => new { e.ValidatorEntityId, e.EpochNumber });
     }
 }