+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/SequencesHolder.cs
@@ -113,4 +113,6 @@ internal class SequencesHolder
     public long PackageCodeHistorySequence { get; set; }
 
     public long PackageSchemaHistorySequence { get; set; }
+
+    public long ValidatorEmissionsSequence { get; set; }
 }